pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hood-sash-automation"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "A package to control a fume hood sash actuator and sensor."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["RPi.GPIO", "Flask", "PyYAML"]

[project.optional-dependencies]
sensor = []
actuator = ["smbus2"]
test = ["pytest", "pytest-mock"]

[project.urls]
"Homepage" = "https://github.com/your-username/fume-hood-sash-automation"
"Bug Tracker" = "https://github.com/your-username/fume-hood-sash-automation/issues"

[project.scripts]
hood_sash_automation_actuator = "hood_sash_automation.actuator.api_service:main"
hood_sash_automation_sensor = "hood_sash_automation.sensor.api_service:main"
